cambio de componente SearchBar
